big update keymap
swap sym and nav layer keys
I found that when I was typing symbols my right thumb would
be swapping back and forth to add a space and was starting to
get tired. Hopefully, this change can relieve this
also, hopefully I can get used to the new positions quickly.
Typing z,q,j,b and shifted versions of those should also be easier
as nav on right and shift on left is more ergonomic.
Also, I won't need shift on symbols nearly as much as I need it
dedicated on nav layer